Implementing modules and generating build library.
Summary
- Initial version of pumascript refactoring into modules
- Using requirejs to generate build pumascript library for browser
and node context
- Updating tests and grunt task
- Moving src/libs/ to thirdparty/ folder
- Adding src/pumascript-dev as a requirejs wrapper to pumascript
main library only for development, to avoid building source code on
each source change.
