+++ package.json
@@ -20,6 +20,7 @@
     "grunt-contrib-clean": "~0.6.0",
     "grunt-contrib-jshint": "~0.11.2",
     "grunt-contrib-nodeunit": "~0.4.1",
+    "grunt-contrib-requirejs": "0.4.4",
     "grunt": "~0.4.5",
     "jshint-stylish": "~1.0.1",
     "load-grunt-tasks": "~3.1.0"
