+++ src/pumascript-run.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8" />
     <title>PumaScript</title>
     <meta name="viewport" content="width=device-width" />
-    <script data-main="pumascript.js" src="./libs/requirejs/require.js"></script>
+    <script data-main="pumascript-dev.js" src="../thirdparty/requirejs/require.js"></script>
    </head>
   <body></body>
 </html>
